Interrupts work so far, some debug thingys
Fixed some bugs caused by typos and brain farts, mainly bad data
types and things I had forgotten. Now, the IDT and GDT seem to wrok,
so far at least. We'll see how it goes when we enter user mode and
start paging.

Some debug things added too: Bochs is now configured to use gdb
(just remove the # from the beginning of the line in bochsrc).
See README for usage.

+++ src/kern/descriptor_tables.h
@@ -41,7 +41,7 @@ struct idt_entry_t
 struct idt_ptr_t
 {
     uint16_t limit; /* Max size of the IDT */
-    uint16_t base;  /* Address of the first element */
+    uint32_t base;  /* Address of the first element */
 }__attribute__((__packed__));
 
 void init_descr_tables();
